+++ 2.7.x/choicemaker-cm/choicemaker-j2ee/com.choicemaker.cm.io.blocking.automated.offline.server/src/main/java/com/choicemaker/cm/io/blocking/automated/offline/server/impl/RecordIdLongTranslation.java
@@ -44,7 +44,7 @@ public class RecordIdLongTranslation extends
 	}
 
 	public RecordIdLongTranslation(BatchJob job, long recordId,
-			RECORD_SOURCE_ROLE source, long translatedId) {
+			RECORD_SOURCE_ROLE source, int translatedId) {
 		super(job.getId(), idToString(recordId), RECORD_ID_TYPE.TYPE_LONG
 				.getCharSymbol(), source.getCharSymbol(), translatedId);
 	}
